// src/interfaces/problem.ts
export interface Example {
  id: number;
  inputText: string;
  outputText: string;
  explanation?: string;
  img?: string;
}

export interface Problem {
  id: string;
  title: string;
  problemStatement: string;
  examples: Example[];
  constraints: string;
  order: number;
  starterCode: string;
  handlerFunction: ((fn: any) => boolean) | string;
  starterFunctionName: string;
}

export interface DBProblem {
  id: string;
  title: string;
  category: string;
  difficulty: string;
  likes: number;
  dislikes: number;
  order: number;
  videoId?: string;
  link?: string;
}